requirement.txt
alembic==1.4.1
click==7.1.1
entrypoints==0.3
fastapi==0.52.0
flake8==3.7.9
h11==0.9.0
httptools==0.1.1
Mako==1.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pydantic==1.4
pyflakes==2.1.1
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
six==1.14.0
SQLAlchemy==1.3.14
starlette==0.13.2
uvicorn==0.11.3
uvloop==0.14.0
websockets==8.1
